Position of light source is not omnipotent (see details)
Earlier a light source was applied to each block equally as an angle.
So if a block was to the left and one to the right of the light source
both would have light on the left side instead of the middles. Now this
problem is fixed such that each block is lighted relative to its
position to the light source.